Modal service

+++ src/app/contact/message-list/message-list.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 
 import { Message } from '../message';
 import { MessagesService } from '../messages.service';
+import { ModalService } from '../../util/modal/modal.service';
+import { MessageConfirmRemoveComponent } from './message-confirm-remove.component';
+import { ModalConfiguration } from '../../util/modal/modal-configuration';
 
 @Component({
     selector: 'message-list',
@@ -13,8 +16,16 @@ import { MessagesService } from '../messages.service';
 export class MessageListComponent implements OnInit {
 
     private messages: Array<Message> = [];
+    private modalConfig: ModalConfiguration = <ModalConfiguration>{
+        title: 'Remove message',
+        cancelBtnEnabled: true,
+        cancelBtnText: 'Cancel',
+        confirmBtnEnabled: true,
+        confirmBtnText: 'Remove'
+    };
 
-    constructor(private messageService: MessagesService) {}
+    constructor(private messageService: MessagesService,
+                private modalService: ModalService) {}
 
     ngOnInit() {
         this.messageService
@@ -25,6 +36,15 @@ export class MessageListComponent implements OnInit {
     }
 
     remove(message: Message): void {
-        this.messageService.remove(message);
+
+        const subscription = this.modalService.open(this.modalConfig, MessageConfirmRemoveComponent);
+
+        subscription
+            .subscribe((response: boolean) => {
+                if (response) {
+                    this.messageService.remove(message);
+                }
+            });
+
     }
 }
